Change api design
