+++ design/api_definition.go
@@ -21,16 +21,15 @@ var _ = API("goa2-sample", func() {
 	})
 
 	cors.Origin("/.*localhost.*/", func() {
-		cors.Headers("authorization, content-type")
 		cors.Methods("GET", "POST", "PUT", "DELETE", "OPTIONS")
 		cors.MaxAge(600)
 	})
 
 	Server("goa2-sample", func() {
-		Services("users", "swagger", "secured", "viron", "stats")
+		Services("Users", "Viron", "Admin")
 		Host("localhost", func() {
 			Description("development host")
-			URI("http://localhost:8000")
+			URI("http://localhost:8080")
 		})
 	})
 })
